♻️ Refactor: 코드레빗 반영
- 삭제 api에서 에러핸들코드 추가
- 카드에서  별점, 리뷰 추가
- 체험 기본 이미지 추가
- 안쓰는 코드 정리

+++ src/features/my/my-activities/lib/api/myActivities.api.ts
@@ -19,6 +19,9 @@ export const getMyActivities = async (
 
 // 내 체험 삭제 api
 export const deleteActivities = async (activityId: number) => {
+  if (!activityId || activityId <= 0) {
+    throw new Error('유효하지 않은 활동 Id입니다.');
+  }
   const response = await api.delete(`/my-activities/${activityId}`);
   return response.data;
 };
